backend/requirements.txt
Flask>=2.0
flask-cors>=3.0
groq
python-dotenv>=0.19